Rename reply interface
