+++ lib/auth.js
@@ -92,7 +92,7 @@ internals.Auth.prototype.test = async function (name, request) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    const response = await request.server._replier.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
+    const response = await request.server._responder.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
 
     if (!response.isAuth) {
         throw response;
@@ -263,7 +263,7 @@ internals.Auth.payload = async function (request) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    const response = await request.server._replier.execute(strategy.methods.payload, request, { bind, realm });
+    const response = await request.server._responder.execute(strategy.methods.payload, request, { bind, realm });
 
     if (response &&
         response.isBoom &&
@@ -294,7 +294,7 @@ internals.Auth.response = async function (request) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    return await request.server._replier.execute(strategy.methods.response, request, { bind, realm, continue: 'undefined' });
+    return await request.server._responder.execute(strategy.methods.response, request, { bind, realm, continue: 'undefined' });
 };
 
 
@@ -318,7 +318,7 @@ internals.authenticate = async function (config, request, manager) {
 
         const bind = strategy.methods;
         const realm = strategy.realm;
-        const response = await request.server._replier.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
+        const response = await request.server._responder.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
 
         const message = (response.isAuth ? internals.validate(response.error, response.data, name, config, request, errors) : internals.validate(response, null, name, config, request, errors));
         if (!message) {
