+++ lib/handler.js
@@ -47,7 +47,7 @@ internals.handler = async function (request, method, pre) {
 
     const bind = request.route.settings.bind;
     const realm = request.route.realm;
-    const response = await request.server._replier.execute(method, request, { bind, realm, continue: 'null' });
+    const response = await request.server._responder.execute(method, request, { bind, realm, continue: 'null' });
 
     // Handler
