+++ lib/plugin.js
@@ -264,7 +264,7 @@ internals.Plugin.prototype.decorate = function (type, property, method, options)
     // Responder
 
     if (type === 'responder') {
-        this.root._replier.decorate(property, method);
+        this.root._responder.decorate(property, method);
         this.root.decorations.responder.push(property);
         return;
     }
