+++ lib/request.js
@@ -405,7 +405,7 @@ internals.Request = class {
             }
 
             if (response &&
-                response !== this.server._replier.continue) {
+                response !== this.server._responder.continue) {
 
                 this._setResponse(response);
 
@@ -451,12 +451,12 @@ internals.Request = class {
             const ext = event.nodes[i];
             const bind = (ext.bind || ext.plugin.realm.settings.bind);
             const realm = ext.plugin.realm;
-            const response = await this.server._replier.execute(ext.func, this, { bind, realm });
+            const response = await this.server._responder.execute(ext.func, this, { bind, realm });
 
             // Process response
 
             if (typeof response === 'symbol') {
-                if (response === this.server._replier.continue) {
+                if (response === this.server._responder.continue) {
                     continue;
                 }
 
@@ -517,7 +517,7 @@ internals.Request = class {
 
     _abort() {
 
-        if (this.response === this.server._replier.close) {
+        if (this.response === this.server._responder.close) {
             this.raw.res.end();                                     // End the response in case it wasn't already closed
         }
