+++ lib/responder.js
