+++ lib/route.js
@@ -387,7 +387,7 @@ internals.payload = async function (request) {
 
         // Custom handler
 
-        return await request.server._replier.execute(failAction, request, { realm: request.route.realm, args: [err] });
+        return await request.server._responder.execute(failAction, request, { realm: request.route.realm, args: [err] });
     }
 };
