+++ lib/server.js
@@ -24,7 +24,7 @@ const Defaults = require('./defaults');
 const Ext = require('./ext');
 const Methods = require('./methods');
 const Plugin = require('./plugin');
-const Reply = require('./reply');
+const Responder = require('./responder');
 const Request = require('./request');
 const Route = require('./route');
 const Schema = require('./schema');
@@ -68,7 +68,7 @@ exports = module.exports = internals.Server = function (options) {
     this._heavy = new Heavy(this._settings.load);
     this._mime = new Mimos(this._settings.mime);
     this._registrations = {};                                                       // Tracks plugin for dependency validation { name -> { version } }
-    this._replier = new Reply();
+    this._responder = new Responder();
     this._requestor = new Request();
     this._plugins = {};                                                             // Exposed plugin properties by name
     this._app = {};
