+++ lib/validation.js
@@ -143,7 +143,7 @@ internals.input = async function (source, request) {
 
     // Custom handler
 
-    return request.server._replier.execute(request.route.settings.validate.failAction, request, { realm: request.route.realm, args: [source, error] });
+    return request.server._responder.execute(request.route.settings.validate.failAction, request, { realm: request.route.realm, args: [source, error] });
 };
 
 
@@ -234,6 +234,6 @@ exports.response = async function (request) {
 
         // Custom handler
 
-        return request.server._replier.execute(request.route.settings.response.failAction, request, { realm: request.route.realm, args: [err] });
+        return request.server._responder.execute(request.route.settings.response.failAction, request, { realm: request.route.realm, args: [err] });
     }
 };
